requirements.txt
openai>=1.30.5,<2
requests>=2.31.0,<3
PyYAML>=6.0.1,<7
google-api-python-client>=2.130.0,<3
google-auth-httplib2>=0.2.0,<1
google-auth-oauthlib>=1.2.0,<2
